password-hash: move trait definitions to toplevel
This change puts the traits front-and-center in the library's
implementation, which is what you'd expect for a crate in the "traits"
repository
